% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunspotsFunctions.R
\name{dtDownloadCompleteSilsoSunspotsTimeSeries}
\alias{dtDownloadCompleteSilsoSunspotsTimeSeries}
\title{Fetch the SILSO sunspots time series}
\usage{
dtDownloadCompleteSilsoSunspotsTimeSeries()
}
\value{
data.table with the following columns:
1) "date_year", 2) "date_month", 3) "date_day", 4) "date_yearfraction",
5) "sunspot_number", 6) "sunspot_number_std", 7) "number_of_observations_used",
8) "is_value_definitive"
}
\description{
Fetch full raw daily SILSO sunspots time series
}
\details{
This function fetches the sunspots data published by SILSO at
http://sidc.be/silso/home and prepares the full time series of all the
observations that have been made available, including for the current month.
}
