% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SunspotsFunctions.R
\name{dtDownloadCurrentMonthSilsoSunspotsTimeSeries}
\alias{dtDownloadCurrentMonthSilsoSunspotsTimeSeries}
\title{Fetch the SILSO sunspots daily time series for current month}
\usage{
dtDownloadCurrentMonthSilsoSunspotsTimeSeries()
}
\value{
data.table in case the fetch is successful; NULL otherwise; The returned
data.table has following columns whose names are largely self-explanatory:
1) "date_year", 2) "date_month", 3) "date_day", 4) "date_yearfraction",
5) "sunspot_number", 6) "sunspot_number_std", 7) "number_of_observations_used",
8) "is_value_definitive"
}
\description{
Fetch the daily SILSO sunspots time series for the current month (updated daily)
}
\details{
This function fetches the time series of sunspots published by SILSO for the
current month. The specific file from which the data is fetched is:

It is stressed that the returned data is raw (e.g. the rows with missing
observations are not excluded, etc.)
}
