% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeomagneticIndicesFunctions.R
\name{dtDownloadGeomagneticIndicesWithSunspotsTimeSeriesForYear}
\alias{dtDownloadGeomagneticIndicesWithSunspotsTimeSeriesForYear}
\title{Download GFZ Potsdam daily geomagnetic, flux and sunspots data}
\usage{
dtDownloadGeomagneticIndicesWithSunspotsTimeSeriesForYear(iYear)
}
\arguments{
\item{iYear}{- integer scalar, not lower thant 1932 (scope of the data set).
This parameter indicates year for which the data is to be downloaded.}
}
\value{
data.table with the downloaded data
}
\description{
Download the daily time series of observations of geomagnetic indices Kp and ap,
solar radio flux and sunspots published by GFZ Potsdam
}
\details{
This function downloads the time series of observations of:
1) magnetic indices Kp and ap
2) international sunspot number
3) solar radio flux
performed by GFZ Potsdam and published by them. More information is
provided on the website https://www.gfz-potsdam.de/en/kp-index/
The files are fetched using the FTP protocol.
This function downloads the data from one of the files "Kp_ap_SN_F107_YYYY.txt"
}
